Added measureFor.
MeasureFor uses timing thread which executes waiter and when waiter
completes it sets a boolean to true which causes measureFor to complete
and return the RunningStat for the loops it run on body.

Eventually there should probably be a pool of waiters so we aren't
continually creating and destroying threads.