Add the beginnings of command line processing.
Discovered the Pegs Module, works but seems limiting. There is a maximum
size of the number of matches of 20 and it doesn't return the number of
elements captured. The information is available in Capture.ml but that
field nor the Capture.matches field are public.

I may suggest that and also allow you to construct a Capture with an
openArray.