Explore asm.
I'm now more comfortable asm and less frustrated, probably the most
important things I learned was the default for inline assembly was att
i.e. "-masm=att" but the code examples in the Intel document "How to
Benchmark Code Execution Times on Intel IA-32 and IA-64 Instruction Set
Architectures" is for -masm=intel, not surprising.

The other significant feature of inline asm was that its counter productive
to explicitly storing results in the inline asm as most of the time they
just degenerate into NOPs i.e. "mov %%eax, %%eax" so it doesn't appear to
be necessary.

Also, with nim it just coding cpuid(), rdtsc() and rdtscp() directly as
inline nim procs generates exactly the same code as coding it as inline
assembler and the the "tuple" variants aren't necessary at all when
doing release builds. I'll simplify bmtool removing the extra stuff
shortly.

Finally, I renamed cmdArgsTable and cmdOptsTable to cmdArgs and cmdOpts
as it shorter and to the point. Overall this way of getting arguments
and options from the command line was fairly convenient.