Rename calibrate to doBmCyclesCalibration and get it working.
Overall templates that return values don't make good templates because
you can't use the "special" syntax where the body can come after a colon.
For example, I'd like to say:

  var rs = doBmCycles 1000:
     nada()

But that doesn't compile instead it has to be:

  var rs = doBmCycles(1000, nada())

An alternative is the "X" templates I created, these pass a var value
as a parameter instead of returning a result. The interesting here is
that I couldn't overload the name but had to have a different name so I
appended an X.